fixing warning refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@9931 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netimport/NIImporter_SUMO.cpp
@@ -134,7 +134,7 @@ NIImporter_SUMO::loadNetwork(const OptionsCont &oc, NBNetBuilder &nb) {
                     if (conn->tlID != "") {
                         NBLoadedSUMOTLDef *tl = (NBLoadedSUMOTLDef*)handler.myTLLCont.getDefinition(conn->tlID);
                         if (tl) {
-                            tl->addConnection(ed->builtEdge, toEdge, fromLaneIndex, toLaneIndex, conn->tlLinkNo);
+                            tl->addConnection(ed->builtEdge, toEdge, fromLaneIndex, (unsigned int)toLaneIndex, conn->tlLinkNo);
                         } else {
                             WRITE_ERROR("The traffic light '" + conn->tlID + "' is not known.");
                         }
